% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearDirectories.R
\name{clearDirectories}
\alias{clearDirectories}
\title{Clear expired directories}
\usage{
clearDirectories(path, reference = NULL, limit = NULL)
}
\arguments{
\item{path}{String containing the path to the base directory, i.e., the directory \emph{containing} all versioned subdirectories for a particular application.}

\item{reference}{A \link{package_version} specifying the most recently accessed version.}

\item{limit}{Integer scalar specifying the maximum number of days to have passed before a versioned directory expires.}
}
\value{
Expired directories are deleted and \code{NULL} is invisibly returned.
}
\description{
Remove versioned directories that have passed on expiration limit.
}
\details{
This function checks the last access date in the \code{*_dir.expiry} files in \code{path}.
If the last access date is too old, the corresponding subdirectory in \code{path} is treated as expired and is deleted.
The age threshold depends on \code{limit}, which defaults to the value of the environment variable \code{BIOC_DIR_EXPIRY_LIMIT}.
If this is not specified, it is set to 30 days.

If \code{reference} is specified, directories with version numbers greater than (or equal to) \code{reference} are deleted,
even if their last access date was older than the specified \code{limit}.
This aims to favor the retention of newer versions, which is generally a sensible outcome when the aim is to stay up-to-date.

This function will acquire an exclusive lock on each directory before attempting to delete it.
Applications can achieve thread safety by calling \code{\link{lockDirectory}} prior to any operations on the versioned directory.
This ensures that \code{clearDirectories} will not delete a directory in use by another process, especially if the latter might update the last access time.
}
\examples{
# Creating the base directory.
base.path <- tempfile(pattern="expired_demo")
dir.create(base.path)

# Creating an older versioned directory.
version <- package_version("1.11.0")
touchDirectory(path, version, date=Sys.Date() - 100, clear=FALSE)
dir.create(file.path(path, version))
list.files(path)

# Clearing them out.
clearDirectories(path)
list.files(path)

}
\seealso{
\code{\link{touchDirectory}}, which calls this function automatically when \code{clear=TRUE}.
}
\author{
Aaron Lun
}
