% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockDirectory.R
\name{lockDirectory}
\alias{lockDirectory}
\title{Lock and unlock directories}
\usage{
lockDirectory(path, version, ...)
}
\arguments{
\item{path}{String containing the path to the base directory, i.e., the directory \emph{containing} all versioned subdirectories for a particular application.}

\item{version}{A \link{package_version} specifying the version to touch.
This should follow the Bioconductor versioning format.}

\item{...}{Further arguments to pass to \code{\link{lock}}.}

\item{handle}{The lock handle generated by \code{\link{lockDirectory}}.}
}
\value{
\code{lockDirectory} returns a lock handle from the \pkg{filelock} package,
applied to a lock file that uses \code{version} as its prefix.

\code{unlockDirectory} unlocks the file referenced by the lock handle.
}
\description{
Mark directories as locked or unlocked for thread-safe processing.
This uses a standard naming scheme for consistency across various applications.
}
\examples{
# Creating the relevant directories.
base.path <- tempfile(pattern="expired_demo")
dir.create(base.path)
version <- package_version("1.11.0")

handle <- lockDirectory(base.path, version)
handle
unlockDirectory(handle)

}
\author{
Aaron Lun
}
