% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockDirectory.R
\name{lockDirectory}
\alias{lockDirectory}
\alias{unlockDirectory}
\title{Lock and unlock directories}
\usage{
lockDirectory(path, version, ...)

unlockDirectory(handle)
}
\arguments{
\item{path}{String containing the path to the base directory, i.e., the directory \emph{containing} all versioned subdirectories for a particular application.}

\item{version}{A \link{package_version} specifying the version to touch.
This should follow the Bioconductor versioning format.}

\item{...}{Further arguments to pass to \code{\link{lock}}.}

\item{handle}{The lock handle generated by \code{\link{lockDirectory}}.}
}
\value{
\code{lockDirectory} returns a pair of lock handles, generated by the \pkg{filelock} package.

\code{unlockDirectory} unlocks the two handles generated by \code{lockDirectory}.
}
\description{
Mark directories as locked or unlocked for thread-safe processing,
using a standard naming scheme for the lock files.
}
\details{
\code{lockDirectory} actually creates two locks.
The first represents a lock on the \code{version} subdirectory within \code{path} and provides thread-safe read/write on its contents.
Concurrent read operations are also permitted by setting \code{exclusive=FALSE} in \code{...} to define a shared lock..

The second lock is applied to \code{path} and is always a shared lock, regardless of the contents of \code{...}.
This provides thread-safe access to the lock file used in the first lock,
protecting it from deletion when the relevant directory expires in \code{\link{clearDirectories}}.
}
\examples{
# Creating the relevant directories.
base.path <- tempfile(pattern="expired_demo")
dir.create(base.path)
version <- package_version("1.11.0")

handle <- lockDirectory(base.path, version)
handle
unlockDirectory(handle)

}
\author{
Aaron Lun
}
