% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touchDirectory.R
\name{touchDirectory}
\alias{touchDirectory}
\title{Touch a versioned directory}
\usage{
touchDirectory(path, version, clear = TRUE, date = Sys.Date(), ...)
}
\arguments{
\item{path}{String containing the path to the base directory, i.e., the directory \emph{containing} all versioned subdirectories for a particular application.}

\item{version}{A \link{package_version} specifying the version to touch.
This should follow the Bioconductor versioning format.}

\item{clear}{Logical scalar indicating whether to remove expired versions.}

\item{date}{A \link{Date} object containing the current date.
Only provided for testing.}

\item{...}{Further arguments to pass to \code{\link{clearDirectories}}.}
}
\value{
The \code{<version>_dir.expiry} stub file within \code{path} is updated/created with the current date.
If \code{clear=TRUE}, expired directories are removed by \code{\link{clearDirectories}}.
A \code{NULL} is invisibly returned.
}
\description{
Touch a versioned directory to indicate that it has been successfully accessed in the recent past.
}
\details{
This function should be called \emph{after} any successful access to the contents of a versioned directory,
to indicate that said directory is still in use by expiry-aware processes.
A stub file is updated with the last access time to allow \code{\link{clearDirectories}} to accurately check for staleness.

For a given \code{path} and \code{version}, this function only modifies the files on its first call.
All subsequent calls with the same two arguments, in the same R session and on the same day will have no effect.
This avoids unnecessary touching of the file system during routine use.

The caller should lock the target directory with \code{\link{lockDirectory}} before attempting access (and certainly before calling this function).
This ensures that another process calling \code{\link{clearDirectories}} does not delete this directory while its access time is being updated.
Under these conditions, any writes to the stub file itself are thread-safe.
}
\examples{
# Creating the base directory.
base.path <- tempfile(pattern="expired_demo")
dir.create(base.path)

# Creating the versioned subdirectory.
version <- package_version("1.11.0")
dir.create(file.path(base.path, version))

# Setting the last access time.
touchDirectory(base.path, version)
list.files(base.path)
readLines(file.path(base.path, "1.11.0_dir.expiry"))

}
\seealso{
\code{\link{clearDirectories}}, to remove expired directories at the same time.
}
\author{
Aaron Lun
}
