% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touchDirectory.R
\name{touchDirectory}
\alias{touchDirectory}
\title{Touch a versioned directory}
\usage{
touchDirectory(path, clear = TRUE, date = Sys.Date(), ...)
}
\arguments{
\item{path}{String containing the path to a versioned directory.
The \code{dirname} should be the package cache while the \code{basename} should be a version number.}

\item{clear}{Logical scalar indicating whether to remove expired versions.}

\item{date}{A \link{Date} object containing the current date.
Only provided for testing.}

\item{...}{Further arguments to pass to \code{\link{clearDirectories}}.}
}
\value{
The \code{<version>_dir.expiry} stub file within \code{path} is updated/created with the current date.
If \code{clear=TRUE}, expired directories are removed by \code{\link{clearDirectories}}.
A \code{NULL} is invisibly returned.
}
\description{
Touch a versioned directory to indicate that it has been successfully accessed in the recent past.
}
\details{
This function should be called \emph{after} any successful access to the contents of a versioned directory,
to indicate that said directory is still in use by expiry-aware processes.
A stub file is updated with the last access time to allow \code{\link{clearDirectories}} to accurately check for staleness.

For a given \code{path} and \code{version}, this function only modifies the files on its first call.
All subsequent calls with the same two arguments, in the same R session and on the same day will have no effect.
This avoids unnecessary touching of the file system during routine use.

The caller should lock the target directory with \code{\link{lockDirectory}} before calling this function.
This ensures that another process calling \code{\link{clearDirectories}} does not delete this directory while its access time is being updated.
If the target directory is locked, any writes to the stub file itself are thread-safe, even for shared locks.
}
\examples{
# Creating the package cache.
cache.dir <- tempfile(pattern="expired_demo")
dir.create(cache.dir)

# Creating the versioned subdirectory.
version <- package_version("1.11.0")
dir.create(file.path(cache.dir, version))

# Setting the last access time.
touchDirectory(file.path(cache.dir, version))
list.files(cache.dir)
readLines(file.path(cache.dir, "1.11.0_dir.expiry"))

}
\seealso{
\code{\link{clearDirectories}}, to remove expired directories at the same time.
}
\author{
Aaron Lun
}
